requirements.txt
aiogram==3.20.0.post0
asyncpg==0.30.0
asyncio==3.4.3
bs4==0.0.2
ebooklib==0.19
fastapi==0.115.14
gigachat==0.1.39.post2
hf-xet==1.1.5
httpx==0.28.1
nltk==3.9.1
numpy==2.3.1
PyMuPDF==1.26.1
pypandoc==1.15
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.4
sentence-transformers==4.1.0
sqlalchemy==2.0.41
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
wikipedia-api==0.8.1
uvicorn==0.35.0
yadisk==3.3.0
